some change, bags clear codes

+++ controllers/posterController.js
@@ -6,12 +6,18 @@ const User = require('../models/userModel')
 // @access     Public
 const getPosterPage = async (req,res) => {
     try{
+
+        // if(req.query.search){
+        //     const searchResoult = await Poster.find({title: req.query.search})
+        //     console.log(searchResoult)
+        // }
+
         const posters = await Poster.find().lean()
         res.render('poster/posters', {
             title: 'Poster Page',
             url: process.env.URL,
-            posters: posters.reverse()
-            
+            posters: posters.reverse(),
+            user: req.session.user
         })
     }catch(err){
         console.log(err)
@@ -32,7 +38,7 @@ const getOnePosterPage = async (req,res) => {
             title: poster.title,
             url: process.env.URL, 
             user: req.session.user, 
-            author: poster.author,
+            author: poster.author, 
             poster
         })
     }catch(err){
@@ -40,47 +46,51 @@ const getOnePosterPage = async (req,res) => {
     }
 }
 
-
 // @Route      GET  /posters/add 
 // @Desc       get add poster page
 // @access     Privide
 const addNewPosterPage = (req,res) => {
-    res.render('poster/addPoster', {
-        title: 'Yangi elon qo\'shish',
-        url: process.env.URL,
-        user: req.session.user,
-        isLogin: req.session.isLogin,
-    })
+    if(req.session.isLogin){
+        res.render('poster/addPoster', {
+            title: 'Yangi elon qo\'shish',
+            url: process.env.URL,
+            user: req.session.user,
+            isLogin: req.session.isLogin,
+            user: req.session.user
+        })
+    }
 }
 
 // @Route      POST  /posters/add
 // @Desc       Add new Poster
 // @access     Privide
 const addNewPoster = async (req,res) => {
-    try{
-        const newPoster = new Poster({
-            title: req.body.title,
-            amount: req.body.amount,
-            region: req.body.region,
-            description: req.body.description,
-            image: 'uploads/' + req.file.filename,
-            author: req.session.user._id
-        })
+    if(req.session.isLogin){
+        try{
+            const newPoster = new Poster({
+                title: req.body.title,
+                amount: req.body.amount,
+                region: req.body.region,
+                description: req.body.description,
+                image: 'uploads/' + req.file.filename,
+                author: req.session.user._id
+            })
 
-        await User.findByIdAndUpdate(req.session.user._id, {
-            $push: {posters: newPoster},
-            new: true,
-            upsert: true 
-        })
-        await newPoster.save((err, posterSaved) => {
-            if(err) throw err
-            const posterId = posterSaved._id
-            res.redirect('/posters/' + posterId)  
-        })
-        
-        
-    }catch (err) { 
-        console.log(err) 
+            await User.findByIdAndUpdate(req.session.user._id, {
+                $push: {posters: newPoster},
+                new: true,
+                upsert: true 
+            })
+            await newPoster.save((err, posterSaved) => {
+                if(err) throw err
+                const posterId = posterSaved._id
+                res.redirect('/posters/' + posterId)  
+            })
+            
+            
+        }catch (err) { 
+            console.log(err) 
+        }
     }
 }
 
@@ -89,47 +99,48 @@ const addNewPoster = async (req,res) => {
 // @access     Privide (own)
 
 const getEditPosterPage = async (req,res) => {
-    try{
-        const poster = await Poster.findById(req.params.id).lean()
-        res.render('poster/editPoster', {
-            title: 'Edit Page',
-            url: process.env.URL,
-            poster
-        })
-    }catch(err){
-        console.log(err)
+    if(req.session.isLogin){
+        try{
+            const poster = await Poster.findById(req.params.id).lean()
+            res.render('poster/editPoster', {
+                title: 'Edit Page',
+                url: process.env.URL,
+                poster,
+                user: req.session.user
+            })
+        }catch(err){
+            console.log(err)
+        }
     }
-    
 }
 
 // @Route      POST  /posters/:id/edit
 // @Desc       Get Edit Poster 
 // @access     Privide (own)
 const updatePoster = async (req,res) => {
-    try{
-        const editedPoster = {
-            title: req.body.title,
-            amount: req.body.amount,
-            image: 'uploads/' + req.file.filename,
-            region: req.body.region,
-            description: req.body.description
+    if(req.session.isLogin){
+        try{
+            const editedPoster = {
+                title: req.body.title,
+                amount: req.body.amount,
+                image: 'uploads/' + req.file.filename,
+                region: req.body.region,
+                description: req.body.description
+            }
+            await Poster.findByIdAndUpdate(req.params.id, editedPoster)
+            res.redirect('/auth/login') 
         }
-        await Poster.findByIdAndUpdate(req.params.id, editedPoster)
-        res.redirect('/posters') 
-    }
-    catch(err){
-        /*if(err == 'undefined'){
-            console.log('salom')
-        }*/
-        const editedPoster = {
-            title: req.body.title,
-            amount: req.body.amount,
-            image:  req.body.image,
-            region: req.body.region,
-            description: req.body.description
+        catch(err){
+            const editedPoster = {
+                title: req.body.title,
+                amount: req.body.amount,
+                image:  req.body.image,
+                region: req.body.region,
+                description: req.body.description
+            }
+            await Poster.findByIdAndUpdate(req.params.id, editedPoster)
+            res.redirect('/posters')
         }
-        await Poster.findByIdAndUpdate(req.params.id, editedPoster)
-        res.redirect('/posters')
     }
 }
 
@@ -137,12 +148,14 @@ const updatePoster = async (req,res) => {
 // @Desc       delete poster by id
 // @access     Privide (own)
 const deletePoster = async (req, res) => {
-    try{
-        await Poster.findByIdAndRemove(req.params.id)
-        res.redirect('/posters')
-    }
-    catch(err){
-        console.log(err)
+    if(req.session.isLogin){
+        try{
+            await Poster.findByIdAndRemove(req.params.id)
+            res.redirect('/posters')
+        }
+        catch(err){
+            console.log(err)
+        }
     }
 }
